refactor: major page simplification and new TableFilter component
**Project Page**: 1064 → 246 lines (76% reduction)
- Extracted 10+ sub-components to _components.tsx
- Added LucideIcon type safety for icons
- Clean orchestration in main page component

**Domains Page**: 575 → 547 lines
- Already well-organized with sub-components
- Minor formatting improvements

**New TableFilter Component**:
- Reusable search/filter UI pattern
- Debounced search input
- Optional filter dropdown and add button
- Export via @/components/dashboard

**Component Library Growth**:
- Added JSDoc documentation to 5 components
- CodeBlock: 77 lines, InlineForm: 116 lines
- TableFilter: 168 lines (new)

**Net Impact**: -627 lines (40% reduction) across modified files
