+++ packages/dashboard/src/components/dashboard/index.ts
@@ -15,3 +15,5 @@ export {
   TableSkeleton,
 } from "./loading-states";
 export { PageHeader } from "./page-header";
+export type { FilterOption } from "./table-filter";
+export { TableFilter } from "./table-filter";
